requirements.txt
amqp==5.1.1
asgiref==3.7.2
asttokens==2.2.1
async-timeout==4.0.3
attrs==23.1.0
Automat==22.10.0
Babel==2.14.0
backcall==0.2.0
beautifulsoup4==4.12.2
billiard==3.6.4.0
bootstrap4==0.1.0
celery==5.2.7
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
colorama==0.4.6
constantly==15.1.0
coreapi==2.3.3
coreschema==0.0.4
crispy-bootstrap5==0.7
cron-descriptor==1.4.0
cryptography==42.0.5
cssselect==1.2.0
decorator==5.1.1
defusedxml==0.8.0rc2
distlib==0.3.7
Django==4.2.3
django-admin-persian-fonts==0.2
django-bootstrap4==23.2
django-celery-beat==2.5.0
django-ckeditor==6.5.1
django-cors-headers==4.3.1
django-crispy-forms==2.0
django-jalali==6.0.1
django-jalali-date==1.1.0
django-js-asset==2.0.0
django-rest-framework==0.1.0
django-rosetta==0.9.8
django-templated-mail==1.1.1
django-timezone-field==6.0
django-tinymce==3.7.1
django-widget-tweaks==1.4.12
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
djoser==2.2.2
docxcompose==1.4.0
docxtpl==0.16.8
drf-yasg==1.21.7
exceptiongroup==1.1.2
executing==1.2.0
filelock==3.12.4
ghasedakpack==0.1.13
h11==0.14.0
hyperlink==21.0.0
idna==3.6
incremental==22.10.0
inflection==0.5.1
ipython==8.13.2
itemadapter==0.8.0
itemloaders==1.1.0
itypes==1.2.0
jdatetime==4.1.1
jedi==0.18.2
Jinja2==3.1.2
jmespath==1.0.1
kombu==5.2.4
lxml==4.9.2
MarkupSafe==2.1.3
matplotlib-inline==0.1.6
mysql-connector-python==8.0.32
mysqlclient==2.2.4
oauthlib==3.2.2
outcome==1.2.0
packaging==23.2
parsel==1.8.1
parso==0.8.3
pickleshare==0.7.5
pillow==10.2.0
pipenv==2023.12.1
platformdirs==3.10.0
polib==1.1.1
prompt-toolkit==3.0.38
Protego==0.2.1
protobuf==3.20.3
psycopg2==2.9.9
pure-eval==0.2.2
pyasn1==0.5.0
pyasn1-modules==0.3.0
pycparser==2.21
PyDispatcher==2.0.7
pyfiglet==1.0.2
pygame==2.5.2
Pygments==2.15.1
pyjokes==0.6.0
PyJWT==2.8.0
pyOpenSSL==23.2.0
PySocks==1.7.1
python-crontab==3.0.0
python-dateutil==2.8.2
python-decouple==3.8
python-docx==1.1.0
python3-openid==3.2.0
pytz==2024.1
PyYAML==6.0.1
queuelib==1.6.2
redis==5.0.3
requests==2.31.0
requests-file==1.5.1
requests-oauthlib==1.3.1
rosetta==0.3
ruamel.yaml==0.17.31
ruamel.yaml.clib==0.2.7
Scrapy==2.9.0
scrapy-djangoitem==1.1.1
scrapyd-api==0.0.1
selenium==4.11.2
service-identity==21.1.0
session-request==0.0.4
six==1.16.0
sniffio==1.3.0
social-auth-app-django==5.4.0
social-auth-core==4.5.3
sortedcontainers==2.4.0
soupsieve==2.4.1
sqlparse==0.4.4
stack-data==0.6.2
tldextract==3.4.4
traitlets==5.9.0
trio==0.22.2
trio-websocket==0.10.3
Twisted==22.10.0
twisted-iocpsupport==1.0.3
typing_extensions==4.6.3
tzdata==2024.1
uritemplate==4.1.1
urllib3==2.2.1
vine==5.0.0
virtualenv==20.24.5
w3lib==2.1.1
wcwidth==0.2.6
wsproto==1.2.0
zope.interface==6.0
